requirements.txt
Discord.py
Asyncio
openpyxl==3.0.9